Backend/akash/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pymongo==4.6.0
supabase==2.3.0
langchain==0.1.0
langchain-community==0.0.10
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0
